+++ RELEASES.md
@@ -8,8 +8,22 @@ Version 1.27.0 (unreleased)
 ===========================
 
   * The [*az* crate], which is a dependency, is now re-exported ([issue 73]).
+  * The following methods were added to [`Float`][flo-1-27]:
+     * [`remainder_quo31`][flo-rq-1-27], [`remainder_quo31_mut`][flo-rqm-1-27],
+       [`remainder_quo31_round`][flo-rqr-1-27] ([issue 76])
+     * [`remainder_quo31_from`][flo-rqf-1-27],
+       [`remainder_quo31_from_round`][flo-rqfr-1-27]
+     * [`remainder_quo31_ref`][flo-rqref-1-27]
 
+[flo-1-27]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html
+[flo-rq-1-27]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html#method.remainder_quo31
+[flo-rqf-1-27]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html#method.remainder_quo31_from
+[flo-rqfr-1-27]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html#method.remainder_quo31_from_round
+[flo-rqm-1-27]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html#method.remainder_quo31_mut
+[flo-rqr-1-27]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html#method.remainder_quo31_round
+[flo-rqref-1-27]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html#method.remainder_quo31_ref
 [issue 73]: https://gitlab.com/tspiteri/rug/-/issues/73
+[issue 76]: https://gitlab.com/tspiteri/rug/-/issues/76
 
 Version 1.26.1 (2024-09-11)
 ===========================