+++ RELEASES.md
@@ -4,6 +4,16 @@
 permitted in any medium without royalty provided the copyright notice and this
 notice are preserved. This file is offered as-is, without any warranty. -->
 
+Version 1.28.0 (unreleased)
+===========================
+
+  * The following methods were added to [`Float`][flo-1-28]:
+     * [`as_shl`][flo-ashl-1-28], [`as_shr`][flo-ashr-1-28]
+
+[flo-1-28]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html
+[flo-ashl-1-28]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html#method.as_shl
+[flo-ashr-1-28]: https://tspiteri.gitlab.io/rug/dev/rug/struct.Float.html#method.as_shr
+
 Version 1.27.0 (2025-01-24)
 ===========================